most of the basic refactoring changes, no logic was changed only refactors, and change in comments
