+++ app/src/main/java/com/github/se/wanderpals/model/data/Trip.kt
@@ -23,8 +23,8 @@ import java.time.LocalDate
  * @param suggestions A list of document IDs referencing the suggestions sub-collection. These
  *   suggestions are potential stops or activities proposed by users for consideration to be
  *   included in the trip itinerary.
- * @param tripNotifications A list of document IDs referencing the tripNotifications sub-collection.
- *   They are the notification linked to a particular trip
+ * @param announcements A list of document IDs referencing the Announcements sub-collection.
+ *   They are the Announcement linked to a particular trip
  */
 data class Trip(
     val tripId: String,
@@ -38,5 +38,5 @@ data class Trip(
     val stops: List<String> = emptyList(),
     val users: List<String> = emptyList(),
     val suggestions: List<String> = emptyList(),
-    val tripNotifications: List<String> = emptyList()
+    val announcements: List<String> = emptyList()
 )
