--- app/src/main/java/com/github/se/wanderpals/model/data/TripNotification.kt
@@ -1,26 +1,0 @@
-package com.github.se.wanderpals.model.data
-
-import java.time.LocalDateTime
-
-/**
- * Represents a notification related to a user's trip within the application. Notifications can be
- * triggered by users or the system itself based on specific conditions or events.
- *
- * @param notificationId The unique identifier for the notification, usually generated by the
- *   database.
- * @param userId The identifier of the user or system that posted the notification. If it's a system
- *   notification, this could represent a system user or null if not applicable.
- * @param title A brief title of the notification.
- * @param userName The name of the user or system entity that posted the notification.
- * @param description Detailed description of the notification, elaborating on the reason or details
- *   of the notification.
- * @param timestamp The exact date and time when the notification was created.
- */
-data class TripNotification(
-    val notificationId: String,
-    val userId: String,
-    val title: String,
-    val userName: String,
-    val description: String,
-    val timestamp: LocalDateTime
-)
