+++ app/src/main/java/com/github/se/wanderpals/model/firestoreData/FirestoreTrip.kt
@@ -8,7 +8,6 @@ import java.time.format.DateTimeFormatter
  * LocalDate fields to String for serialization. It supports conversion between the Trip domain
  * model and the FirestoreTrip DTO to facilitate easy data handling with Firestore.
  *
- * @param tripNotifications List of IDs for trip notifications.
  * @property tripId Unique identifier of the trip, usually generated by the database.
  * @property title Name or title providing a brief overview of the trip.
  * @property startDate Start date of the trip in the format (yyyy-MM-dd).
@@ -19,6 +18,7 @@ import java.time.format.DateTimeFormatter
  * @property stops List of IDs for the trip's stops.
  * @property users List of participant user IDs.
  * @property suggestions List of IDs for suggested activities or stops.
+ * @property announcements List of IDs for trip announcements.
  */
 data class FirestoreTrip(
     val tripId: String = "",
@@ -31,7 +31,7 @@ data class FirestoreTrip(
     val stops: List<String> = emptyList(),
     val users: List<String> = emptyList(),
     val suggestions: List<String> = emptyList(),
-    val tripNotifications: List<String> = emptyList()
+    val announcements: List<String> = emptyList()
 ) {
   companion object {
     /**
@@ -53,7 +53,7 @@ data class FirestoreTrip(
           stops = trip.stops,
           users = trip.users,
           suggestions = trip.suggestions,
-          tripNotifications = trip.tripNotifications)
+          announcements = trip.announcements)
     }
   }
 
@@ -75,6 +75,6 @@ data class FirestoreTrip(
         stops = stops,
         users = users,
         suggestions = suggestions,
-        tripNotifications = tripNotifications)
+        announcements = announcements)
   }
 }
