+++ app/src/main/java/com/github/se/wanderpals/model/repository/TripsRepository.kt
@@ -6,11 +6,11 @@ import com.github.se.wanderpals.model.data.Role
 import com.github.se.wanderpals.model.data.Stop
 import com.github.se.wanderpals.model.data.Suggestion
 import com.github.se.wanderpals.model.data.Trip
-import com.github.se.wanderpals.model.data.TripNotification
+import com.github.se.wanderpals.model.data.Announcement
 import com.github.se.wanderpals.model.data.User
 import com.github.se.wanderpals.model.firestoreData.FirestoreStop
 import com.github.se.wanderpals.model.firestoreData.FirestoreSuggestion
-import com.github.se.wanderpals.model.firestoreData.FirestoreTripNotification
+import com.github.se.wanderpals.model.firestoreData.FirestoreAnnouncement
 import com.github.se.wanderpals.model.firestoreData.FirestoreUser
 import com.github.se.wanderpals.service.SessionManager
 import com.google.firebase.FirebaseApp
@@ -69,65 +69,65 @@ class TripsRepository(
   }
 
   /**
-   * Retrieves a specific trip notification from a trip based on the notification's unique
-   * identifier. This method queries the Firestore subcollection for trip notifications within a
+   * Retrieves a specific trip Announcement from a trip based on the Announcement's unique
+   * identifier. This method queries the Firestore subcollection for trip Announcements within a
    * specific trip document.
    *
    * @param tripId The unique identifier of the trip.
-   * @param tripNotificationId The unique identifier of the trip notification to be retrieved.
-   * @return A `TripNotification` object if found, or `null` if the notification is not found or if
+   * @param announcementId The unique identifier of the trip Announcement to be retrieved.
+   * @return A `Announcement` object if found, or `null` if the Announcement is not found or if
    *   an error occurs. The method logs an error and returns `null` in case of failure.
    */
-  suspend fun getTripNotificationFromTrip(
+  suspend fun getAnnouncementFromTrip(
       tripId: String,
-      tripNotificationId: String
-  ): TripNotification? =
+      announcementId: String
+  ): Announcement? =
       withContext(dispatcher) {
         try {
           val documentSnapshot =
               tripsCollection
                   .document(tripId)
                   .collection(FirebaseCollections.TRIPS_NOTIFICATION_SUBCOLLECTION.path)
-                  .document(tripNotificationId)
+                  .document(announcementId)
                   .get()
                   .await()
-          val firestoreTripNotification = documentSnapshot.toObject<FirestoreTripNotification>()
-          if (firestoreTripNotification != null) {
-            firestoreTripNotification.toTripNotification()
+          val firestoreAnnouncement = documentSnapshot.toObject<FirestoreAnnouncement>()
+          if (firestoreAnnouncement != null) {
+            firestoreAnnouncement.toAnnouncement()
           } else {
             Log.e(
                 "TripsRepository",
-                "getTripNotificationFromTrip: Not found TripNotification $tripNotificationId from trip $tripId.")
+                "getTripNotificationFromTrip: Not found TripNotification $announcementId from trip $tripId.")
             null
           }
         } catch (e: Exception) {
           Log.e(
               "TripsRepository",
-              "getTripNotificationFromTrip: Error getting a TripNotification $tripNotificationId from trip $tripId.",
+              "getTripNotificationFromTrip: Error getting a TripNotification $announcementId from trip $tripId.",
               e)
           null // error
         }
       }
 
   /**
-   * Retrieves all trip notifications associated with a specific trip. It iterates over all
-   * notification IDs stored within the trip document and fetches their corresponding trip
-   * notification objects.
+   * Retrieves all trip Announcements associated with a specific trip. It iterates over all
+   * Announcement IDs stored within the trip document and fetches their corresponding trip
+   * Announcement objects.
    *
    * @param tripId The unique identifier of the trip.
-   * @return A list of `TripNotification` objects. Returns an empty list if the trip is not found,
-   *   if there are no notifications associated with the trip, or in case of an error during data
+   * @return A list of `Announcement` objects. Returns an empty list if the trip is not found,
+   *   if there are no Announcements associated with the trip, or in case of an error during data
    *   retrieval.
    */
-  suspend fun getAllTripNotificationsFromTrip(tripId: String): List<TripNotification> =
+  suspend fun getAllAnnouncementsFromTrip(tripId: String): List<Announcement> =
       withContext(dispatcher) {
         try {
           val trip = getTrip(tripId)
 
           if (trip != null) {
-            val tripNotificationIds = trip.tripNotifications
-            tripNotificationIds.mapNotNull { tripNotificationId ->
-              getTripNotificationFromTrip(tripId, tripNotificationId)
+            val announcementIds = trip.announcements
+            announcementIds.mapNotNull { announcementId ->
+              getAnnouncementFromTrip(tripId, announcementId)
             }
           } else {
             Log.e(
@@ -146,44 +146,44 @@ class TripsRepository(
       }
 
   /**
-   * Adds a notification to a specific trip in the Firestore database. This method generates a
-   * unique ID for the new notification, creates a corresponding FirestoreTripNotification object,
-   * and commits it to the trip's notification subcollection.
+   * Adds a Announcement to a specific trip in the Firestore database. This method generates a
+   * unique ID for the new Announcement, creates a corresponding FirestoreAnnouncement object,
+   * and commits it to the trip's Announcements subcollection.
    *
-   * It also updates the trip's document to include this new notification ID in its list.
+   * It also updates the trip's document to include this new Announcement ID in its list.
    *
-   * @param tripId The unique identifier of the trip where the notification will be added.
-   * @param tripNotification The TripNotification object to add.
+   * @param tripId The unique identifier of the trip where the Announcement will be added.
+   * @param announcement The Announcement object to add.
    * @return Boolean indicating the success (true) or failure (false) of the operation.
    */
-  suspend fun addTripNotificationToTrip(
+  suspend fun addAnnouncementToTrip(
       tripId: String,
-      tripNotification: TripNotification
+      announcement: Announcement
   ): Boolean =
       withContext(dispatcher) {
         try {
           val uniqueID = UUID.randomUUID().toString()
-          val firebaseTripNotification =
-              FirestoreTripNotification.fromTripNotification(
-                  tripNotification.copy(
-                      notificationId = uniqueID,
-                      userId = uid)) // we already know who creates the TripNotification
-          val tripNotificationDocument =
+          val firebaseAnnouncement =
+              FirestoreAnnouncement.fromAnnouncement(
+                  announcement.copy(
+                      announcementId = uniqueID,
+                      userId = uid)) // we already know who creates the Announcement
+          val announcementDocument =
               tripsCollection
                   .document(tripId)
                   .collection(FirebaseCollections.TRIPS_NOTIFICATION_SUBCOLLECTION.path)
                   .document(uniqueID)
-          tripNotificationDocument.set(firebaseTripNotification).await()
+          announcementDocument.set(firebaseAnnouncement).await()
           Log.d(
               "TripsRepository",
               "addTripNotificationToTrip: TripNotification added successfully to trip $tripId.")
 
           val trip = getTrip(tripId)
           if (trip != null) {
-            // Add the new tripNotificationId to the trip's tripsNotifications list and update the
+            // Add the new AnnouncementId to the trip's Announcements list and update the
             // trip
-            val updatedTripNotificationsList = trip.tripNotifications + uniqueID
-            val updatedTrip = trip.copy(tripNotifications = updatedTripNotificationsList)
+            val updatedAnnouncementsList = trip.announcements + uniqueID
+            val updatedTrip = trip.copy(announcements = updatedAnnouncementsList)
             updateTrip(updatedTrip)
             Log.d(
                 "TripsRepository",
@@ -204,37 +204,37 @@ class TripsRepository(
       }
 
   /**
-   * Removes a specific notification from a trip's notification list and the Firestore database.
-   * This method first deletes the notification document from the trip's notification subcollection.
+   * Removes a specific Announcement from a trip's Announcement list and the Firestore database.
+   * This method first deletes the Announcement document from the trip's Announcement subcollection.
    *
-   * It also updates the trip's document to remove the notification ID from its list.
+   * It also updates the trip's document to remove the Announcement ID from its list.
    *
    * @param tripId The unique identifier of the trip.
-   * @param tripNotificationId The unique identifier of the notification to be removed.
+   * @param announcementId The unique identifier of the Announcement to be removed.
    * @return Boolean indicating the success (true) or failure (false) of the operation.
    */
-  suspend fun removeTripNotificationFromTrip(tripId: String, tripNotificationId: String): Boolean =
+  suspend fun removeTripNotificationFromTrip(tripId: String, announcementId: String): Boolean =
       withContext(dispatcher) {
         try {
           Log.d(
               "TripsRepository",
-              "removeSuggestionFromTrip: removing TripNotification $tripNotificationId from trip $tripId")
+              "removeSuggestionFromTrip: removing TripNotification $announcementId from trip $tripId")
           tripsCollection
               .document(tripId)
               .collection(FirebaseCollections.TRIPS_NOTIFICATION_SUBCOLLECTION.path)
-              .document(tripNotificationId)
+              .document(announcementId)
               .delete()
               .await()
 
           val trip = getTrip(tripId)
           if (trip != null) {
-            val updatedTripNotificationsList =
-                trip.tripNotifications.filterNot { it == tripNotificationId }
-            val updatedTrip = trip.copy(suggestions = updatedTripNotificationsList)
+            val updatedAnnouncementsList =
+                trip.announcements.filterNot { it == announcementId }
+            val updatedTrip = trip.copy(announcements = updatedAnnouncementsList)
             updateTrip(updatedTrip)
             Log.d(
                 "TripsRepository",
-                "removeSuggestionFromTrip: TripNotification $tripNotificationId remove and trip updated successfully.")
+                "removeSuggestionFromTrip: TripNotification $announcementId remove and trip updated successfully.")
             true
           } else {
             Log.e("TripsRepository", "removeSuggestionFromTrip: Trip not found with ID $tripId.")
@@ -243,38 +243,38 @@ class TripsRepository(
         } catch (e: Exception) {
           Log.e(
               "TripsRepository",
-              "removeSuggestionFromTrip: Error removing TripNotification $tripNotificationId from trip $tripId.",
+              "removeSuggestionFromTrip: Error removing TripNotification $announcementId from trip $tripId.",
               e)
           false
         }
       }
 
   /**
-   * Updates a specific notification in a trip's notification subcollection in the Firestore
-   * database. This method creates a FirestoreTripNotification object from the provided
-   * TripNotification object and sets it to the corresponding document identified by the
+   * Updates a specific Announcement in a trip's Announcement subcollection in the Firestore
+   * database. This method creates a FirestoreAnnouncement object from the provided
+   * Announcement object and sets it to the corresponding document identified by the
    * notification ID.
    *
    * @param tripId The unique identifier of the trip.
-   * @param tripNotification The TripNotification object that contains updated data.
+   * @param announcement The Announcement object that contains updated data.
    * @return Boolean indicating the success (true) or failure (false) of the operation.
    */
   suspend fun updateTripNotificationInTrip(
       tripId: String,
-      tripNotification: TripNotification
+      announcement: Announcement
   ): Boolean =
       withContext(dispatcher) {
         try {
           Log.d(
               "TripsRepository",
               "updateTripNotificationInTrip: Updating a TripNotification in trip $tripId")
-          val firestoreTripNotification =
-              FirestoreTripNotification.fromTripNotification(tripNotification)
+          val firestoreAnnouncement =
+              FirestoreAnnouncement.fromAnnouncement(announcement)
           tripsCollection
               .document(tripId)
               .collection(FirebaseCollections.TRIPS_NOTIFICATION_SUBCOLLECTION.path)
-              .document(firestoreTripNotification.notificationId)
-              .set(firestoreTripNotification)
+              .document(firestoreAnnouncement.announcementId)
+              .set(firestoreAnnouncement)
               .await()
           Log.d(
               "TripsRepository",
@@ -283,7 +283,7 @@ class TripsRepository(
         } catch (e: Exception) {
           Log.e(
               "TripsRepository",
-              "updateTripNotificationInTrip: Error updating tripNotification with ID ${tripNotification.notificationId} in trip with ID $tripId",
+              "updateTripNotificationInTrip: Error updating tripNotification with ID ${announcement.announcementId} in trip with ID $tripId",
               e)
           false
         }
