+++ app/src/main/java/com/github/se/wanderpals/model/viewmodel/NotificationViewModel.kt
@@ -1,6 +1,6 @@
 package com.github.se.wanderpals.model.viewmodel
 
-import com.github.se.wanderpals.model.data.TripNotification
+import com.github.se.wanderpals.model.data.Announcement
 import com.github.se.wanderpals.model.repository.TripsRepository
 import com.github.se.wanderpals.service.SessionManager
 
@@ -13,11 +13,11 @@ class NotificationViewModel(private val tripsRepository: TripsRepository) {
    * @param tripNotification The notification object to be deleted.
    * @return A boolean representing the success of the operation.
    */
-  suspend fun deleteTripNotification(tripId: String, tripNotification: TripNotification): Boolean {
+  suspend fun deleteTripNotification(tripId: String, tripNotification: Announcement): Boolean {
     // Check if the current user has permission to remove items, only admins can delete
     // notifications
     return if (SessionManager.canRemove(tripNotification.userId)) {
-      tripsRepository.removeTripNotificationFromTrip(tripId, tripNotification.notificationId)
+      tripsRepository.removeTripNotificationFromTrip(tripId, tripNotification.announcementId)
     } else {
       false
     }
