+++ app/src/test/java/com/github/se/wanderpals/firestoreData/FirestoreTripNotificationTest.kt
@@ -1,56 +1,56 @@
 package com.github.se.wanderpals.model.firestoreData
 
-import com.github.se.wanderpals.model.data.TripNotification
+import com.github.se.wanderpals.model.data.Announcement
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
 
-class TripNotificationTest {
+class AnnouncementTest {
 
   private val formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME
 
   @Test
-  fun testFromTripNotificationToFirestoreTripNotification() {
-    val tripNotification =
-        TripNotification(
-            notificationId = "notif123",
+  fun testFromAnnouncementToFirestoreAnnouncement() {
+    val announcement =
+        Announcement(
+            announcementId = "notif123",
             userId = "user456",
             title = "Trip Alert",
             userName = "Jane Doe",
             description = "Your trip to Paris is coming up soon!",
             timestamp = LocalDateTime.parse("2024-04-01T12:30:00", formatter))
 
-    val firestoreTripNotification = FirestoreTripNotification.fromTripNotification(tripNotification)
+    val firestoreAnnouncement = FirestoreAnnouncement.fromAnnouncement(announcement)
 
-    assertEquals(tripNotification.notificationId, firestoreTripNotification.notificationId)
-    assertEquals(tripNotification.userId, firestoreTripNotification.userId)
-    assertEquals(tripNotification.title, firestoreTripNotification.title)
-    assertEquals(tripNotification.userName, firestoreTripNotification.userName)
-    assertEquals(tripNotification.description, firestoreTripNotification.description)
-    assertEquals(tripNotification.timestamp.format(formatter), firestoreTripNotification.timestamp)
+    assertEquals(announcement.announcementId, firestoreAnnouncement.announcementId)
+    assertEquals(announcement.userId, firestoreAnnouncement.userId)
+    assertEquals(announcement.title, firestoreAnnouncement.title)
+    assertEquals(announcement.userName, firestoreAnnouncement.userName)
+    assertEquals(announcement.description, firestoreAnnouncement.description)
+    assertEquals(announcement.timestamp.format(formatter), firestoreAnnouncement.timestamp)
   }
 
   @Test
-  fun testFromFirestoreTripNotificationToTripNotification() {
-    val firestoreTripNotification =
-        FirestoreTripNotification(
-            notificationId = "notif123",
+  fun testFromFirestoreAnnouncementToAnnouncement() {
+    val firestoreAnnouncement =
+        FirestoreAnnouncement(
+            announcementId = "notif123",
             userId = "user456",
             title = "Trip Alert",
             userName = "Jane Doe",
             description = "Your trip to Paris is coming up soon!",
             timestamp = "2024-04-01T12:30:00")
 
-    val tripNotification = firestoreTripNotification.toTripNotification()
+    val announcement = firestoreAnnouncement.toAnnouncement()
 
-    assertEquals(firestoreTripNotification.notificationId, tripNotification.notificationId)
-    assertEquals(firestoreTripNotification.userId, tripNotification.userId)
-    assertEquals(firestoreTripNotification.title, tripNotification.title)
-    assertEquals(firestoreTripNotification.userName, tripNotification.userName)
-    assertEquals(firestoreTripNotification.description, tripNotification.description)
+    assertEquals(firestoreAnnouncement.announcementId, announcement.announcementId)
+    assertEquals(firestoreAnnouncement.userId, announcement.userId)
+    assertEquals(firestoreAnnouncement.title, announcement.title)
+    assertEquals(firestoreAnnouncement.userName, announcement.userName)
+    assertEquals(firestoreAnnouncement.description, announcement.description)
     assertEquals(
-        LocalDateTime.parse(firestoreTripNotification.timestamp, formatter),
-        tripNotification.timestamp)
+        LocalDateTime.parse(firestoreAnnouncement.timestamp, formatter),
+        announcement.timestamp)
   }
 }
